# src/app/dto/app.dto.go
package app

type CreateAppDto struct {
	Sample string`json:"sample,omitempty"`
}

type UpdateAppDto struct {
	Sample string`json:"sample,omitempty"`
}

type QueryAppDto struct {
	Sample string`json:"sample,omitempty"`
}